frontend/requirements.txt
Flask==1.1.2
requests==2.25.0
python-dotenv==0.15.0